requirements.txt
# ============================================
# Spore AI Agent - Python Dependencies
# ============================================

# ------------------------------------------
# Core LLM SDK
# ------------------------------------------
openai>=1.51.0              # OpenAI / DeepSeek API
anthropic>=0.40.0           # Claude API
httpx>=0.27.0               # HTTP client (required by anthropic)

# ------------------------------------------
# Configuration & Environment
# ------------------------------------------
python-dotenv>=1.0.1        # .env file loading
PyYAML>=6.0.0               # YAML configuration parsing

# ------------------------------------------
# Token Counting
# ------------------------------------------
tiktoken>=0.7.0             # GPT/DeepSeek token counting

# ------------------------------------------
# Terminal & Clipboard
# ------------------------------------------
colorama>=0.4.6             # Terminal colored output
pyperclip>=1.8.0            # Clipboard operations

# ------------------------------------------
# Windows System Support
# ------------------------------------------
pywin32>=306; sys_platform == 'win32'  # Windows named pipes (win32pipe, win32file, pywintypes)

# ------------------------------------------
# Document Processing
# ------------------------------------------
docxtpl>=0.16.7             # Word document templates
python-docx>=0.8.11         # Word document manipulation

# ------------------------------------------
# Web Scraping & Search
# ------------------------------------------
requests>=2.31.0            # HTTP requests
beautifulsoup4>=4.12.0      # HTML parsing (bs4)
lxml>=4.9.0                 # XML/HTML parser (faster than html.parser)
ddgs>=0.6.0                 # DuckDuckGo search (ddgs metasearch)

# ------------------------------------------
# Desktop Mode (FastAPI + Tauri)
# ------------------------------------------
fastapi>=0.109.0            # REST API framework
uvicorn>=0.27.0             # ASGI server
websockets>=12.0            # WebSocket support
pydantic>=2.0.0             # Data validation

# ------------------------------------------
# Web Service Mode (Flask) - Optional
# ------------------------------------------
# flask>=3.0.0              # Web framework
# flask-cors>=4.0.0         # CORS support
# flask-socketio>=5.3.0     # Socket.IO support
# eventlet>=0.33.0          # Async networking

# ------------------------------------------
# System Monitoring (Optional)
# ------------------------------------------
psutil>=5.9.0               # Process/system information

# ------------------------------------------
# Development & Testing (Optional)
# ------------------------------------------
# pytest>=8.0.0             # Testing framework
# hypothesis>=6.100.0       # Property-based testing

# ------------------------------------------
# Build Tools (Required for packaging)
# ------------------------------------------
pyinstaller>=6.0.0          # Package Python backend into executable

# ------------------------------------------
# Binary Analysis (Optional - for IDA-Skill)
# ------------------------------------------
# pefile>=2023.2.7          # PE file parser
# pycryptodome>=3.19.0      # Cryptography library (AES, RSA, etc.)

# ============================================
# External Tool Dependencies (Install Separately)
# ============================================
# 
# 1. ripgrep (rg) - Fast file search
#    Windows: choco install ripgrep
#    macOS:   brew install ripgrep
#    Linux:   apt install ripgrep
#
# 2. Node.js (18.x or 20.x LTS) + npm - Frontend build
#    Download: https://nodejs.org/
#    Verify: node --version && npm --version
#
# 3. Rust + Cargo - Tauri desktop app
#    Windows: https://rustup.rs/
#    Verify: rustc --version && cargo --version
#
# 4. Visual Studio Build Tools (Windows only)
#    Required for Tauri compilation
#    Download: https://visualstudio.microsoft.com/downloads/
#    Components: MSVC v143, Windows 10/11 SDK
#
# 5. NSIS (Optional) - Windows installer creation
#    Windows: choco install nsis
#    Download: https://nsis.sourceforge.io/
#
# 6. UPX (Optional) - Executable compression
#    Windows: choco install upx
#    Download: https://upx.github.io/
#
# ============================================
# Frontend Dependencies (npm packages)
# ============================================
# See: desktop_app/frontend/package.json
#
# Production:
#   @tauri-apps/api: ^1.5.3
#   react: ^18.2.0
#   react-dom: ^18.2.0
#   zustand: ^4.4.7
#
# Development:
#   @tauri-apps/cli: ^1.5.9
#   @types/node: ^25.0.3
#   @types/react: ^18.2.43
#   @types/react-dom: ^18.2.17
#   @vitejs/plugin-react: ^4.2.1
#   autoprefixer: ^10.4.16
#   postcss: ^8.4.32
#   tailwindcss: ^3.4.0
#   typescript: ^5.3.3
#   vite: ^5.0.10
#
# Install: cd desktop_app/frontend && npm install
#
# ============================================
# Build Instructions
# ============================================
# 
# For detailed build instructions, see BUILD_REQUIREMENTS.md
#
# Quick Start:
#   1. Install Python dependencies: pip install -r requirements.txt
#   2. Install frontend dependencies: cd desktop_app/frontend && npm install
#   3. Build desktop app: build_desktop.bat (Windows)
#   4. Build installer: build_installer.bat (Windows)
#
# ============================================
